LPD-45735 buildREST

+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/BaseGroupedProductResourceImpl.java
@@ -432,9 +432,14 @@ public abstract class BaseGroupedProductResourceImpl
 
 		UnsafeFunction<GroupedProduct, GroupedProduct, Exception>
 			groupedProductUnsafeFunction = groupedProduct -> {
-				deleteGroupedProduct(groupedProduct.getId());
+				if (groupedProduct.getId() != null) {
+					deleteGroupedProduct(groupedProduct.getId());
 
-				return groupedProduct;
+					return groupedProduct;
+				}
+
+				throw new UnsupportedOperationException(
+					"Unable to delete groupedProduct. No valid identifier provided.");
 			};
 
 		if (contextBatchUnsafeBiConsumer != null) {
