+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/BaseMappedProductResourceImpl.java
@@ -540,9 +540,14 @@ public abstract class BaseMappedProductResourceImpl
 
 		UnsafeFunction<MappedProduct, MappedProduct, Exception>
 			mappedProductUnsafeFunction = mappedProduct -> {
-				deleteMappedProduct(mappedProduct.getId());
+				if (mappedProduct.getId() != null) {
+					deleteMappedProduct(mappedProduct.getId());
 
-				return mappedProduct;
+					return mappedProduct;
+				}
+
+				throw new UnsupportedOperationException(
+					"Unable to delete mappedProduct. No valid identifier provided.");
 			};
 
 		if (contextBatchUnsafeBiConsumer != null) {
