+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/BaseProductAccountGroupResourceImpl.java
@@ -287,9 +287,14 @@ public abstract class BaseProductAccountGroupResourceImpl
 
 		UnsafeFunction<ProductAccountGroup, ProductAccountGroup, Exception>
 			productAccountGroupUnsafeFunction = productAccountGroup -> {
-				deleteProductAccountGroup(productAccountGroup.getId());
+				if (productAccountGroup.getId() != null) {
+					deleteProductAccountGroup(productAccountGroup.getId());
 
-				return productAccountGroup;
+					return productAccountGroup;
+				}
+
+				throw new UnsupportedOperationException(
+					"Unable to delete productAccountGroup. No valid identifier provided.");
 			};
 
 		if (contextBatchUnsafeBiConsumer != null) {
