+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/BaseProductChannelResourceImpl.java
@@ -277,9 +277,14 @@ public abstract class BaseProductChannelResourceImpl
 
 		UnsafeFunction<ProductChannel, ProductChannel, Exception>
 			productChannelUnsafeFunction = productChannel -> {
-				deleteProductChannel(productChannel.getId());
+				if (productChannel.getId() != null) {
+					deleteProductChannel(productChannel.getId());
 
-				return productChannel;
+					return productChannel;
+				}
+
+				throw new UnsupportedOperationException(
+					"Unable to delete productChannel. No valid identifier provided.");
 			};
 
 		if (contextBatchUnsafeBiConsumer != null) {
