+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/BaseProductConfigurationListAccountGroupResourceImpl.java
@@ -465,11 +465,19 @@ public abstract class BaseProductConfigurationListAccountGroupResourceImpl
 			 ProductConfigurationListAccountGroup, Exception>
 				productConfigurationListAccountGroupUnsafeFunction =
 					productConfigurationListAccountGroup -> {
-						deleteProductConfigurationListAccountGroup(
-							productConfigurationListAccountGroup.
-								getProductConfigurationListAccountGroupId());
+						if (productConfigurationListAccountGroup.
+								getProductConfigurationListAccountGroupId() !=
+									null) {
 
-						return productConfigurationListAccountGroup;
+							deleteProductConfigurationListAccountGroup(
+								productConfigurationListAccountGroup.
+									getProductConfigurationListAccountGroupId());
+
+							return productConfigurationListAccountGroup;
+						}
+
+						throw new UnsupportedOperationException(
+							"Unable to delete productConfigurationListAccountGroup. No valid identifier provided.");
 					};
 
 		if (contextBatchUnsafeBiConsumer != null) {
