+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/BaseProductConfigurationListAccountResourceImpl.java
@@ -460,11 +460,18 @@ public abstract class BaseProductConfigurationListAccountResourceImpl
 			<ProductConfigurationListAccount, ProductConfigurationListAccount,
 			 Exception> productConfigurationListAccountUnsafeFunction =
 				productConfigurationListAccount -> {
-					deleteProductConfigurationListAccount(
-						productConfigurationListAccount.
-							getProductConfigurationListAccountId());
+					if (productConfigurationListAccount.
+							getProductConfigurationListAccountId() != null) {
 
-					return productConfigurationListAccount;
+						deleteProductConfigurationListAccount(
+							productConfigurationListAccount.
+								getProductConfigurationListAccountId());
+
+						return productConfigurationListAccount;
+					}
+
+					throw new UnsupportedOperationException(
+						"Unable to delete productConfigurationListAccount. No valid identifier provided.");
 				};
 
 		if (contextBatchUnsafeBiConsumer != null) {
