+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/BaseProductConfigurationListOrderTypeResourceImpl.java
@@ -464,11 +464,19 @@ public abstract class BaseProductConfigurationListOrderTypeResourceImpl
 			 ProductConfigurationListOrderType, Exception>
 				productConfigurationListOrderTypeUnsafeFunction =
 					productConfigurationListOrderType -> {
-						deleteProductConfigurationListOrderType(
-							productConfigurationListOrderType.
-								getProductConfigurationListOrderTypeId());
+						if (productConfigurationListOrderType.
+								getProductConfigurationListOrderTypeId() !=
+									null) {
 
-						return productConfigurationListOrderType;
+							deleteProductConfigurationListOrderType(
+								productConfigurationListOrderType.
+									getProductConfigurationListOrderTypeId());
+
+							return productConfigurationListOrderType;
+						}
+
+						throw new UnsupportedOperationException(
+							"Unable to delete productConfigurationListOrderType. No valid identifier provided.");
 					};
 
 		if (contextBatchUnsafeBiConsumer != null) {
