+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/BaseProductGroupProductResourceImpl.java
@@ -414,9 +414,14 @@ public abstract class BaseProductGroupProductResourceImpl
 
 		UnsafeFunction<ProductGroupProduct, ProductGroupProduct, Exception>
 			productGroupProductUnsafeFunction = productGroupProduct -> {
-				deleteProductGroupProduct(productGroupProduct.getId());
+				if (productGroupProduct.getId() != null) {
+					deleteProductGroupProduct(productGroupProduct.getId());
 
-				return productGroupProduct;
+					return productGroupProduct;
+				}
+
+				throw new UnsupportedOperationException(
+					"Unable to delete productGroupProduct. No valid identifier provided.");
 			};
 
 		if (contextBatchUnsafeBiConsumer != null) {
