+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/BaseProductOptionValueResourceImpl.java
@@ -360,9 +360,14 @@ public abstract class BaseProductOptionValueResourceImpl
 
 		UnsafeFunction<ProductOptionValue, ProductOptionValue, Exception>
 			productOptionValueUnsafeFunction = productOptionValue -> {
-				deleteProductOptionValue(productOptionValue.getId());
+				if (productOptionValue.getId() != null) {
+					deleteProductOptionValue(productOptionValue.getId());
 
-				return productOptionValue;
+					return productOptionValue;
+				}
+
+				throw new UnsupportedOperationException(
+					"Unable to delete productOptionValue. No valid identifier provided.");
 			};
 
 		if (contextBatchUnsafeBiConsumer != null) {
