+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/BaseProductVirtualSettingsFileEntryResourceImpl.java
@@ -334,10 +334,15 @@ public abstract class BaseProductVirtualSettingsFileEntryResourceImpl
 			<ProductVirtualSettingsFileEntry, ProductVirtualSettingsFileEntry,
 			 Exception> productVirtualSettingsFileEntryUnsafeFunction =
 				productVirtualSettingsFileEntry -> {
-					deleteProductVirtualSettingsFileEntry(
-						productVirtualSettingsFileEntry.getId());
+					if (productVirtualSettingsFileEntry.getId() != null) {
+						deleteProductVirtualSettingsFileEntry(
+							productVirtualSettingsFileEntry.getId());
 
-					return productVirtualSettingsFileEntry;
+						return productVirtualSettingsFileEntry;
+					}
+
+					throw new UnsupportedOperationException(
+						"Unable to delete productVirtualSettingsFileEntry. No valid identifier provided.");
 				};
 
 		if (contextBatchUnsafeBiConsumer != null) {
