+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/BaseRelatedProductResourceImpl.java
@@ -448,9 +448,14 @@ public abstract class BaseRelatedProductResourceImpl
 
 		UnsafeFunction<RelatedProduct, RelatedProduct, Exception>
 			relatedProductUnsafeFunction = relatedProduct -> {
-				deleteRelatedProduct(relatedProduct.getId());
+				if (relatedProduct.getId() != null) {
+					deleteRelatedProduct(relatedProduct.getId());
 
-				return relatedProduct;
+					return relatedProduct;
+				}
+
+				throw new UnsupportedOperationException(
+					"Unable to delete relatedProduct. No valid identifier provided.");
 			};
 
 		if (contextBatchUnsafeBiConsumer != null) {
