+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-channel-impl/src/main/java/com/liferay/headless/commerce/admin/channel/internal/resource/v1_0/BaseChannelAccountResourceImpl.java
@@ -371,9 +371,14 @@ public abstract class BaseChannelAccountResourceImpl
 
 		UnsafeFunction<ChannelAccount, ChannelAccount, Exception>
 			channelAccountUnsafeFunction = channelAccount -> {
-				deleteChannelAccount(channelAccount.getChannelAccountId());
+				if (channelAccount.getChannelAccountId() != null) {
+					deleteChannelAccount(channelAccount.getChannelAccountId());
 
-				return channelAccount;
+					return channelAccount;
+				}
+
+				throw new UnsupportedOperationException(
+					"Unable to delete channelAccount. No valid identifier provided.");
 			};
 
 		if (contextBatchUnsafeBiConsumer != null) {
