+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-channel-impl/src/main/java/com/liferay/headless/commerce/admin/channel/internal/resource/v1_0/BaseProductDisplayPageResourceImpl.java
@@ -456,9 +456,14 @@ public abstract class BaseProductDisplayPageResourceImpl
 
 		UnsafeFunction<ProductDisplayPage, ProductDisplayPage, Exception>
 			productDisplayPageUnsafeFunction = productDisplayPage -> {
-				deleteProductDisplayPage(productDisplayPage.getId());
+				if (productDisplayPage.getId() != null) {
+					deleteProductDisplayPage(productDisplayPage.getId());
 
-				return productDisplayPage;
+					return productDisplayPage;
+				}
+
+				throw new UnsupportedOperationException(
+					"Unable to delete productDisplayPage. No valid identifier provided.");
 			};
 
 		if (contextBatchUnsafeBiConsumer != null) {
