+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-order-impl/src/main/java/com/liferay/headless/commerce/admin/order/internal/resource/v1_0/BaseOrderRuleAccountResourceImpl.java
@@ -410,10 +410,15 @@ public abstract class BaseOrderRuleAccountResourceImpl
 
 		UnsafeFunction<OrderRuleAccount, OrderRuleAccount, Exception>
 			orderRuleAccountUnsafeFunction = orderRuleAccount -> {
-				deleteOrderRuleAccount(
-					orderRuleAccount.getOrderRuleAccountId());
+				if (orderRuleAccount.getOrderRuleAccountId() != null) {
+					deleteOrderRuleAccount(
+						orderRuleAccount.getOrderRuleAccountId());
 
-				return orderRuleAccount;
+					return orderRuleAccount;
+				}
+
+				throw new UnsupportedOperationException(
+					"Unable to delete orderRuleAccount. No valid identifier provided.");
 			};
 
 		if (contextBatchUnsafeBiConsumer != null) {
