+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-pricing-impl/src/main/java/com/liferay/headless/commerce/admin/pricing/internal/resource/v1_0/BaseDiscountProductResourceImpl.java
@@ -401,9 +401,14 @@ public abstract class BaseDiscountProductResourceImpl
 
 		UnsafeFunction<DiscountProduct, DiscountProduct, Exception>
 			discountProductUnsafeFunction = discountProduct -> {
-				deleteDiscountProduct(discountProduct.getId());
+				if (discountProduct.getId() != null) {
+					deleteDiscountProduct(discountProduct.getId());
 
-				return discountProduct;
+					return discountProduct;
+				}
+
+				throw new UnsupportedOperationException(
+					"Unable to delete discountProduct. No valid identifier provided.");
 			};
 
 		if (contextBatchUnsafeBiConsumer != null) {
